Add album view
Now to play music...
Also, the performance is questionable. When build with optimizations its
okay, but without its not. + even with optimizations it sometimes has a
noticeable delay when reading an album, which we don't want. I gotta
figure out how to load content in the background and display it as it
updates.

Signed-off-by: Kyle Robert Dawson <kylerdawson@proton.me>